WIP: Started fixing the BlockCollection spec
Unfortunately there are a couple of things that need to be done: Since we'll be storing the IDs in the header, we'll need to think of how we want to handle a header initially being empty. Do we want the BlockCollection to be responsible for fetching the data, or do we want something higher above handling it?